Changed watch size model to decimalfield

+++ products/models.py
@@ -9,7 +9,7 @@ class Product(models.Model):
     watch_make = models.CharField(max_length=254)
     watch_model = models.CharField(max_length=254)
     description = models.TextField()
-    watch_size = models.IntegerField()
+    watch_size = models.DecimalField(max_digits=2, decimal_places=0)
     price = models.DecimalField(max_digits=6, decimal_places=3)
     image_url = models.URLField(max_length=1024, null=True, blank=True)
     image = models.ImageField(null=True, blank=True)
